Issue #79: Consolidate and remove redundant parser tests - closes #79
Move unique location tracking tests from parser/tests.rs into parser.rs and remove
redundant duplicates. This consolidates tests with their implementations following
standard Rust practice of co-locating tests with the code they test.

Changes:
- Move test_location_tracking_for_core_elements to parser.rs
- Move test_location_tracking_for_annotations to parser.rs
- Move test_location_tracking_for_foreign_blocks to parser.rs
- Remove test_simple_paragraph (duplicate of parser.rs)
- Remove test_real_content_extraction (duplicate of parser.rs)
- Remove test_malformed_session_title_with_indent_but_no_content (duplicate of sessions.rs)
- Remove test_session_title_followed_by_bare_indent_level (duplicate of sessions.rs)
- Delete redundant src/txxt/parser/tests.rs file
- Remove tests module reference from src/txxt/parser.rs

Result: Centralized testing, improved maintainability, reduced code duplication.
All 207 tests passing.
